Allow users to add mountains

+++ client/src/components/adminPanel/AdminMountains.tsx
@@ -1,6 +1,6 @@
 import { useMutation, useQuery } from '@apollo/react-hooks';
 import gql from 'graphql-tag';
-import React, { useState } from 'react';
+import React, { useContext, useState } from 'react';
 import { withRouter } from 'react-router';
 import {
   ContentBody,
@@ -9,17 +9,20 @@ import {
   ContentRightSmall as MountainEditColumn,
 } from '../../styling/Grid';
 import { ButtonPrimary } from '../../styling/styleUtils';
-import { Mountain } from '../../types/graphQLTypes';
+import { Mountain, PermissionTypes } from '../../types/graphQLTypes';
 import { failIfValidOrNonExhaustive } from '../../Utils';
+import { UserContext } from '../App';
 import StandardSearch from '../sharedComponents/StandardSearch';
 import { GET_PEAK_LISTS } from './AdminPeakLists';
 import AddMountain from './mountains/AddMountain';
 import EditMountain from './mountains/EditMountain';
 import ListMountains from './mountains/ListMountains';
+import {
+  NavButtonLink,
+  SubNav,
+} from './sharedStyles';
 
-export const GET_MOUNTAINS = gql`
-  query ListMountains {
-    mountains {
+const getMountainsQuery = `
       id
       name
       latitude
@@ -31,6 +34,35 @@ export const GET_MOUNTAINS = gql`
         name
         abbreviation
       }
+      author {
+        id
+        mountainPermissions
+        email
+      }
+      flag
+      status
+`;
+
+export const GET_MOUNTAINS = gql`
+  query ListMountains {
+    mountains {
+      ${getMountainsQuery}
+    }
+  }
+`;
+
+const GET_FLAGGED_MOUNTAINS = gql`
+  query ListFlaggedMountains {
+    mountains: flaggedMountains {
+      ${getMountainsQuery}
+    }
+  }
+`;
+
+const GET_PENDING_MOUNTAINS = gql`
+  query ListPendingMountains {
+    mountains: pendingMountains {
+      ${getMountainsQuery}
     }
   }
 `;
@@ -43,6 +75,7 @@ const ADD_MOUNTAIN = gql`
       $elevation: Float!,
       $prominence: Float,
       $state: ID!,
+      $author: ID!,
     ) {
     addMountain(
       name: $name,
@@ -51,17 +84,9 @@ const ADD_MOUNTAIN = gql`
       elevation: $elevation,
       prominence: $prominence,
       state: $state,
+      author: $author,
     ) {
-      id
-      name
-      latitude
-      longitude
-      elevation
-      prominence
-      state {
-        id
-        name
-      }
+      ${getMountainsQuery}
     }
   }
 `;
@@ -85,21 +110,12 @@ const EDIT_MOUNTAIN = gql`
       prominence: $prominence,
       state: $state,
     ) {
-      id
-      name
-      latitude
-      longitude
-      elevation
-      prominence
-      state {
-        id
-        name
-      }
+      ${getMountainsQuery}
     }
   }
 `;
 
-export interface AddMountainVariables {
+interface BaseMountainVariables {
   name: string;
   latitude: number;
   longitude: number;
@@ -107,9 +123,14 @@ export interface AddMountainVariables {
   prominence: number | null;
   state: string;
 }
-export type EditMountainVariables = AddMountainVariables & {
+
+export interface AddMountainVariables extends BaseMountainVariables {
+  author: string;
+}
+
+export interface EditMountainVariables extends BaseMountainVariables {
   id: string;
-};
+}
 
 const DELETE_MOUNTAIN = gql`
   mutation($id: ID!) {
@@ -129,8 +150,33 @@ enum EditMountainPanelEnum {
   Update,
 }
 
+enum MountainsToShow {
+  all,
+  pending,
+  flagged,
+}
+
 const AdminMountains = () => {
-  const {loading, error, data} = useQuery<SuccessResponse>(GET_MOUNTAINS);
+
+  const user = useContext(UserContext);
+
+  const [mountainsToShow, setMountainsToShow] = useState<MountainsToShow>(MountainsToShow.all);
+
+  let query;
+  if (mountainsToShow === MountainsToShow.all) {
+    query = GET_MOUNTAINS;
+  } else if (mountainsToShow === MountainsToShow.flagged) {
+    query = GET_FLAGGED_MOUNTAINS;
+  } else if (mountainsToShow === MountainsToShow.pending) {
+    query = GET_PENDING_MOUNTAINS;
+  } else {
+    failIfValidOrNonExhaustive(mountainsToShow,
+      'Invalid type for mountainsToShow ' + mountainsToShow);
+    query = GET_MOUNTAINS;
+  }
+
+  const {loading, error, data} = useQuery<SuccessResponse>(query);
+
   const [editMountainPanel, setEditMountainPanel] = useState<EditMountainPanelEnum>(EditMountainPanelEnum.Empty);
   const [mountainToEdit, setMountainToEdit] = useState<string | null>(null);
   const [searchQuery, setSearchQuery] = useState<string>('');
@@ -213,7 +259,15 @@ const AdminMountains = () => {
   };
 
   const deleteMountain = (id: string) => {
-    deleteMountainMutation({ variables: { id } });
+    const mountainToDelete = data && data.mountains
+      ? data.mountains.find(mtn => mtn.id === id) : undefined;
+    const authorId = mountainToDelete && mountainToDelete.author
+      ? mountainToDelete.author.id : null;
+    if (user &&
+        (user._id === authorId || user.permissions === PermissionTypes.admin)
+       ) {
+      deleteMountainMutation({ variables: { id } });
+    }
     clearEditMountainPanel();
   };
 
@@ -226,6 +280,23 @@ const AdminMountains = () => {
       <MountainListColumn>
         <ContentHeader>
           <h2>Mountains</h2>
+          <SubNav>
+            <NavButtonLink
+              onClick={() => setMountainsToShow(MountainsToShow.all)}
+            >
+              All mountains
+            </NavButtonLink>
+            <NavButtonLink
+              onClick={() => setMountainsToShow(MountainsToShow.pending)}
+            >
+              Pending mountains
+            </NavButtonLink>
+            <NavButtonLink
+              onClick={() => setMountainsToShow(MountainsToShow.flagged)}
+            >
+              Flagged mountains
+            </NavButtonLink>
+          </SubNav>
           <StandardSearch
             placeholder={'Filter mountains'}
             setSearchQuery={filterMountains}
